+++ CHANGELOG.md
@@ -10,6 +10,7 @@ doivent être mises à jour #1667
 - Ajout des permissions daemon:lecture et daemon:edition permettant aux utilisateurs de voir l'état des dossiers en cours #2192
 - Création de la table `configuration` pour stocker les paramètres applicatifs. #2232
 - Possibilité de créer des gestionnaires de tâches spécifiques à chaque entité via l'interface de configuration #2233
+- Ajout des pages relatives à l'accessibilité #1951
 
 ## Évolutions